requirements.txt
streamlit
pandas
requests
plotly
streamlit-autorefresh
twilio
groq
python-dotenv
scikit-learn
faker